Avoid wrong (or should I say none) emulation problem of the div instruction on the Coldfire by casting the source/destination operand u in function PUTL from ulong to long.

+++ sys/libkern/vsprintf.c
@@ -99,7 +99,7 @@ PUTL (char *p, long *cnt, ulong u, int base, int width, int fill_char)
 	
 	do {
 		*t++ = "0123456789ABCDEF"[u % base];
-		u /= base;
+		(long)u /= base;
 		width--;
 	}
 	while (u > 0);
